requirements.txt
pillow
blinker
